+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Accept remote identity proofs created using `eddsa-jcs-2022` cryptosuite.
 
+### Changed
+
+- Check origin of verification method when verifying activity signature.
+
 ## [3.20.0] - 2025-03-15
 
 ### Added